split blueprint into multiple files
