+++ ckanext/unhcr/blueprints/__init__.py
@@ -1,0 +1,2 @@
+from .access_requests import unhcr_access_requests_blueprint
+from .metrics import unhcr_metrics_blueprint
