+++ ckanext/unhcr/templates/user/snippets/access_request_modals.html
@@ -4,7 +4,7 @@
     <div class="modal-content">
 
       <form
-        action="{{ h.url_for('unhcr_access_requests.access_requests_reject', request_id=req.id, qualified=True) }}"
+        action="{{ h.url_for('unhcr_access_requests.reject', request_id=req.id, qualified=True) }}"
         method="POST"
       >
